# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation",
        "title": "My API Swagger",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/meal-log-items/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new meal log item record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal_log_item"
                ],
                "summary": "Create meal log item",
                "parameters": [
                    {
                        "description": "Meal log item data",
                        "name": "meal_log_item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MealLogItem"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Meal log item created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.MealLogItem"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/meal-log-items/food/{foodId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all meal log items associated with a specific food",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal_log_item"
                ],
                "summary": "Get meal log items by food ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Food ID",
                        "name": "foodId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of meal log items for the food",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MealLogItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid food ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/meal-log-items/meal-log/{mealLogId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all items belonging to a specific meal log",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal_log_item"
                ],
                "summary": "Get meal log items by meal log ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal log ID",
                        "name": "mealLogId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of meal log items",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MealLogItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid meal log ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete all meal log items associated with a specific meal log ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal_log_item"
                ],
                "summary": "Delete all items for a meal log",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal log ID",
                        "name": "mealLogId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All meal log items deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid meal log ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/meal-log-items/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific meal log item record by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal_log_item"
                ],
                "summary": "Get meal log item by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal log item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Meal log item retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.MealLogItem"
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Meal log item not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing meal log item by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal_log_item"
                ],
                "summary": "Update meal log item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal log item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated meal log item data",
                        "name": "meal_log_item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MealLogItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Meal log item updated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.MealLogItem"
                        }
                    },
                    "400": {
                        "description": "Invalid ID or request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a specific meal log item by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal_log_item"
                ],
                "summary": "Delete meal log item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal log item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Meal log item deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/meal-logs/{id}/items": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add one or more food items to an existing meal log (ownership required)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meal_log_item"
                ],
                "summary": "Add items to an existing meal log",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Meal log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Items to add",
                        "name": "meal_log_item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddItemsToMealLogRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Items added to meal log successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MealLogItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID or request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden — not the owner",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddItemsToMealLogRequestDTO": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MealLogItemDTO"
                    }
                }
            }
        },
...
        "dto.MealLogItemDTO": {
            "type": "object",
            "properties": {
                "food_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "quantity_grams": {
                    "type": "number"
                }
            }
        },
...
        "models.MealLogItem": {
            "type": "object",
            "properties": {
                "food_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "meal_log_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "quantity_grams": {
                    "type": "number"
                }
            }
        },
...
    }
}